// src/store/api/types.ts
export enum Step {
  address = 'address',
  coverage = 'coverage',
  businessInformation = 'businessInformation',
  checkout = 'checkout',
}

export enum QuoteStatus {
  INFO_COLLECTED = 'INFO_COLLECTED',
  PAYMENT_PENDING = 'PAYMENT_PENDING',
  BOUND = 'BOUND',
}

export interface IPolicy {
  quoteEstimates: IQuoteEstimate[];
  selectedEstimateId: string;
  amount: number;
  effectiveDateUtc: string;
}

export interface IAddress {
  street: string;
  street2: string;
  city: string;
  state: string;
  zipCode: string;
}

export interface ICoverage {
  coverageAmount: number;
  estimateId: string;
  effectiveDate: string;
}
// ...
export interface IQuoteEstimate {
  type: string;
  duration: number;
  productId: string;
  countyFips: string;
  netRateOnline: number;
  premiumAmount: number;
  coverageAmount: number;
}